rewrite registration of routes
